Adding more content and fleshing out the story. Support story events
This commit adds more content and a general story outline now that I've figured
out what that will look like. It also adds redux to add support for story
events, so that when you visit certain areas in the story, it causes an event
that will affect your choices in other areas.

There is a little more story work to flesh out, and some mroe instructional
content and then this should be ready for candidates